Let user select Global Properties
Problem
The build explorer only showed the default configuration and platform

This was a problem if you were trying to inspect build properties for a
different configuration or platform than what was set as the default
in the build project file.

Analysis
This is not a hard problem to solve. The msbuild API allows for setting
arbitrary properties as global variables. Not only that, the msbuild API also
has a nice property that gets all the possible configs and platforms that
currently exist inside of an msbuild file.

Solution
Added two combo boxes to the main ribbon that display the current
configuration and platform for the file. This is front and center to indicate
to the user the importance fact that the properties are interpreted through
the lense of the config and platform and that they can change them

I also cleaned a few things up here.